Sketch out search routines, fix ordering property for Value (#2)
The Value type is an i16 that encodes a score in centipawns as well as
the distance from winning and losing positions. In order to preserve the
property that worse scores compare worse than better scores, this commit
reverses the order of mate scores. This nicely preserves the negation
property while also ensuring that the total order on i16 also works as
the total order of Values.