[python] Add STSTest
The Strategic Test Suite (STS) is a list of positions along with best moves
and scores associated with a number of moves. It's a reasonably simple way
to estimate the strength of an engine over a wide variety of positions. The `sts`
module aims to be a driver for the STS test suite for any UCI compliant engine,
including a4 (and stockfish, if we want to test ourselves).