[moveorder] Considering moving attacked pieces first
Pieces that are attacked (particularly if they are hanging) often
need to be moved. This commit pushes them forward in the move ordering
scheme so that we'll consider moves that move away from threats early
on.

a4 has a tendency to hang pieces so this is an attempt to remedy that.